refactor : 보너스 번호 객체를 제거, 당첨 번호 객체에 보너스 번호 기능 포함
