+++ README.md
@@ -49,12 +49,10 @@
     - [x] 1 ~ 45 사이의 값을 입력 받아 당첨 번호를 생성한다.
         - [x] 당첨 번호가 6개 보다 작거나 많은 경우 예외가 발생한다.
         - [x] 당첨 번호가 중복될 경우 예외가 발생한다.
-    - [x] 당첨 번호와 일치하는 로또 번호 개수를 반환한다.
-
-- 보너스 번호
     - [x] 1 ~ 45 사이의 값으로 보너스 번호를 생성한다.
-        - [x] 당첨 번호와 중복될 경우 예외가 발생한다.
         - [x] 보너스 번호가 1 ~ 45 사이의 값이 아닐 경우 예외가 발생한다.
+        - [x] 당첨 번호와 중복될 경우 예외가 발생한다.
+    - [x] 당첨 번호와 일치하는 로또 번호 개수를 반환한다.
 
 - 당첨 결과
     - [x] 당첨 순위별로 당첨된 로또 티켓 개수를 조회한다.
