+++ src/test/java/lotto/domain/winning/WinningNumbersTest.java
@@ -24,7 +24,7 @@ class WinningNumbersTest {
     void createWinningNumbers() {
         //given //when
         List<Integer> numbers = Arrays.asList(1, 2, 3, 4, 5, 6);
-        WinningNumbers winningNumbers = new WinningNumbers(numbers);
+        WinningNumbers winningNumbers = new WinningNumbers(numbers, 7);
 
         //then
         assertThat(winningNumbers.getWinningNumbers()).hasSize(6);
@@ -41,7 +41,7 @@ class WinningNumbersTest {
 
         //then
         assertThatIllegalArgumentException()
-                .isThrownBy(() -> new WinningNumbers(numbers));
+                .isThrownBy(() -> new WinningNumbers(numbers, 7));
     }
 
     @Test
@@ -52,7 +52,7 @@ class WinningNumbersTest {
 
         //when //then
         assertThatIllegalArgumentException()
-                .isThrownBy(() -> new WinningNumbers(numbers));
+                .isThrownBy(() -> new WinningNumbers(numbers, 7));
     }
 
     @ParameterizedTest(name = "당첨 번호와 일치하는 개수: [{index}] {1}")
@@ -60,7 +60,7 @@ class WinningNumbersTest {
     @DisplayName("당첨 번호와 일치하는 로또 번호 개수를 반환한다.")
     void matchWinningNumbers(LottoTicket lottoTicket, int excepted) {
         //given
-        WinningNumbers winningNumbers = new WinningNumbers(Arrays.asList(1, 2, 3, 4, 5, 6));
+        WinningNumbers winningNumbers = new WinningNumbers(Arrays.asList(1, 2, 3, 4, 5, 6), 7);
 
         //when
         int matchCount = winningNumbers.matchCount(lottoTicket.getLottoNumbers());
@@ -79,5 +79,29 @@ class WinningNumbersTest {
                 Arguments.of(LottoTicket.from(6, 7, 8, 9, 10, 11), 1),
                 Arguments.of(LottoTicket.from(7, 8, 9, 10, 11, 12), 0)
         );
+    }
+
+    @ParameterizedTest
+    @ValueSource(ints = {0, 46})
+    @DisplayName("보너스 번호가 1 ~ 45 사이의 값이 아닐 경우 예외가 발생한다.")
+    void validateRange(int invalidBonusNumber) {
+        //given
+        List<Integer> winningNumbers = Arrays.asList(1, 2, 3, 4, 5, 6);
+
+        //when //then
+        assertThatIllegalArgumentException()
+                .isThrownBy(() -> new WinningNumbers(winningNumbers, invalidBonusNumber));
+    }
+
+    @Test
+    @DisplayName("보너스 번호가 당첨 번호와 중복될 경우 예외가 발생한다.")
+    void validateDuplicationBonus() {
+        //given
+        List<Integer> winningNumbers = Arrays.asList(1, 2, 3, 4, 5, 6);
+        int invalidBonusNumber = 5;
+
+        //when //then
+        assertThatIllegalArgumentException()
+                .isThrownBy(() -> new WinningNumbers(winningNumbers, invalidBonusNumber));
     }
 }
